Remove Gradle template, add minimal Android manifest
Unity doesn't seem to set `applicationId` correctly so `pathPrefix` for
the Auth0 redirect activity is set to a value that doesn't match the
correct redirect url. Seems the only way to set the Auth0 parameters
reliably is to hardcode them in the manifest, without this Android can't
redirect the user from the browser back to Unity.