Resend txs on nonce mismatch
If a tx is rejected by the DAppChain due to a nonce mismatch the client
will attempt to resend it again after 0.5 seconds. By default the client
will make a maximum of 5 attempts to resend a tx before giving up,
this can be adjusted by setting the `NonceRetries` property to a another
value.